# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Netriun Nexus API",
    "version": "1.10.0",
    "description": "Multi-tenant workspace API with live provider refresh, database snapshots, Alibaba ECS security-group rule management, custom account access policies, local verified-email authentication, and workspace-scoped OIDC/SAML SSO. Lists return {data: [...]}; errors return {error: message}. Login tokens expire after 12 hours. Community workspaces include one owner and five additional users. The workspace owner retains local break-glass access when SSO is required."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Sign in; returns token, user, expires_in and session cookie",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Constraint conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        },
        "security": []
      }
    },
    "/auth/signup": {
      "post": {
        "summary": "Create an isolated Community workspace and send owner email verification",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Signup"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Workspace created; email verification is required before sign-in"
          },
          "400": {
            "description": "Invalid request; the response includes `field` for workspace, username, email, or password validation failures"
          },
          "409": {
            "description": "Email or username already in use"
          },
          "429": {
            "description": "Sign-up rate limit exceeded"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": []
      }
    },
    "/auth/resend-verification": {
      "post": {
        "summary": "Request a replacement email-verification link",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": { "email": { "type": "string", "format": "email" } },
                "required": ["email"]
              }
            }
          }
        },
        "responses": {
          "202": { "description": "Generic response returned whether or not the account exists" },
          "429": { "description": "Rate limit exceeded" }
        },
        "security": []
      }
    },
    "/auth/sso/providers": {
      "get": {
        "summary": "Discover enabled identity providers using a workspace SSO code",
        "parameters": [{ "name": "workspace", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Enabled provider names, protocols, and sign-in URLs" },
          "400": { "description": "Workspace code is required" },
          "429": { "description": "Discovery rate limit exceeded" }
        },
        "security": []
      }
    },
    "/auth/me": {
      "get": {
        "summary": "Current user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Constraint conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Revoke current session",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Constraint conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      }
    },
    "/summary": {
      "get": {
        "summary": "Scoped inventory counts",
        "parameters": [
          {
            "name": "account_ids",
            "in": "query",
            "description": "Comma-separated accessible account IDs from one cloud provider",
            "schema": { "type": "string", "pattern": "^[1-9][0-9]*(,[1-9][0-9]*)*$" }
          },
          {
            "name": "account_id",
            "in": "query",
            "deprecated": true,
            "description": "Backward-compatible single-account filter; use account_ids",
            "schema": { "type": "integer", "minimum": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Constraint conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      }
    },
    "/reports/overview": {
      "get": {
        "summary": "Resource consumption and billing-readiness report",
        "description": "Aggregates the latest healthy compute, EDS and OSS snapshots for accessible accounts. Exact spend remains unavailable until provider billing APIs and separate cost-read permissions are configured.",
        "parameters": [
          {
            "name": "account_ids",
            "in": "query",
            "description": "Comma-separated accessible account IDs from one cloud provider",
            "schema": { "type": "string", "pattern": "^[1-9][0-9]*(,[1-9][0-9]*)*$" }
          }
        ],
        "responses": {
          "200": { "description": "Latest snapshot report" },
          "400": { "description": "Invalid account selection" },
          "401": { "description": "Authentication required" },
          "403": { "description": "Cloud account access required" },
          "503": { "description": "Dependency unavailable" }
        }
      }
    },
    "/reports/alibaba/billing": {
      "get": {
        "summary": "Build an exact Alibaba monthly billing report",
        "description": "Reads previously synchronized Alibaba BSS instance-bill rows. Totals use PretaxAmount and remain separated by service and currency.",
        "parameters": [
          { "name": "account_ids", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "billing_cycle", "in": "query", "schema": { "type": "string", "pattern": "^[0-9]{4}-(0[1-9]|1[0-2])$" } },
          { "name": "services", "in": "query", "description": "Comma-separated ecs, eds, or other values", "schema": { "type": "string", "default": "ecs,eds" } },
          { "name": "resource_query", "in": "query", "description": "Case-insensitive resource name or instance ID substring", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Exact stored bill rows and totals" },
          "400": { "description": "Invalid account, month, or service selection" },
          "403": { "description": "billing.view access required" }
        }
      }
    },
    "/reports/alibaba/billing/sync": {
      "post": {
        "summary": "Synchronize exact Alibaba BSS bill rows",
        "description": "Fetches all DescribeInstanceBill pages and atomically replaces the selected account/month snapshots. Requires bssapi:DescribeInstanceBill in RAM and billing.sync in Nexus.",
        "parameters": [
          { "name": "account_ids", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "billing_cycle", "in": "query", "schema": { "type": "string", "pattern": "^[0-9]{4}-(0[1-9]|1[0-2])$" } }
        ],
        "responses": {
          "200": { "description": "Synchronization completed" },
          "400": { "description": "Invalid scope" },
          "403": { "description": "billing.sync access required" },
          "502": { "description": "Alibaba BSS rejected or failed the request" }
        }
      }
    },
    "/reports/alibaba/billing/export": {
      "get": {
        "summary": "Export a filtered Alibaba billing report",
        "description": "Uses the same account, month, service, and resource filters as the report endpoint.",
        "parameters": [
          { "name": "account_ids", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "billing_cycle", "in": "query", "schema": { "type": "string" } },
          { "name": "services", "in": "query", "schema": { "type": "string" } },
          { "name": "resource_query", "in": "query", "schema": { "type": "string" } },
          { "name": "format", "in": "query", "required": true, "schema": { "type": "string", "enum": ["xlsx", "pdf"] } }
        ],
        "responses": {
          "200": { "description": "Excel workbook or PDF document" },
          "400": { "description": "Invalid export parameters" },
          "403": { "description": "billing.view access required" }
        }
      }
    },
    "/accounts": {
      "get": {
        "summary": "List accounts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Constraint conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      },
      "post": {
        "summary": "Create Account",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Constraint conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Account"
              }
            }
          }
        }
      }
    },
    "/accounts/test": {
      "post": {
        "summary": "Validate new cloud credentials and required compute read access",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ConnectionTest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Credentials and required read access are working" },
          "400": { "description": "Invalid provider or credentials" },
          "401": { "description": "Authentication required" },
          "403": { "description": "Cloud account management access required" },
          "429": { "description": "Connection-test rate limit exceeded" },
          "502": { "description": "Provider rejected the connection; response includes a stable NX error code and remediation steps" }
        }
      }
    },
    "/accounts/{id}/test": {
      "post": {
        "summary": "Validate a saved cloud connection or replacement credentials",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "minimum": 1 }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ConnectionTest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Stored or replacement credentials and required read access are working" },
          "400": { "description": "Invalid provider or replacement credentials" },
          "401": { "description": "Authentication required" },
          "403": { "description": "Account management access required" },
          "404": { "description": "Account not found" },
          "429": { "description": "Connection-test rate limit exceeded" },
          "502": { "description": "Provider rejected the connection; response includes a stable NX error code and remediation steps" }
        }
      }
    },
    "/accounts/{id}": {
      "put": {
        "summary": "Update Account",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Constraint conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Account"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Account",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Constraint conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      }
    },
...
    "/activity": {
      "get": {
        "summary": "List activity",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Constraint conflict"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      }
    },
...
    "/accounts/{id}/eds/regions": {
      "get": {
        "summary": "Read the latest WUYING EDS region and desktop-count snapshot",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "minimum": 1 } }
        ],
        "responses": { "200": { "description": "Region snapshot with desktop counts; unavailable counts are -1" }, "400": { "description": "Invalid provider" }, "403": { "description": "Access denied" } }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "nexus_session"
      }
    },
    "schemas": {
...
      "Login": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "ConnectionTest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "provider": {
            "type": "string",
            "enum": ["aws", "alibaba", "azure", "gcp"]
          },
          "credentials": {
            "$ref": "#/components/schemas/Account/properties/credentials"
          }
        },
        "required": ["provider"]
      },
...
    }
  }
}